requirements.txt
streamlit 
python-dotenv 
langchain_google_genai
langchain
langchain_community
psycopg2-binary